Inline existing menus options
